requirements.txt
torch 
torchvision 
flask 
flask-socketio 
eventlet 
opencv-python 
numpy 
pillow
pylint